requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
imbalanced-learn==0.11.0
shap==0.43.0
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.1.0
